rename some stuff

+++ app/src/main/java/covidproject/Data.java
@@ -65,7 +65,7 @@ public class Data {
 			while ((rowString = reader.readLine()) != null) {
 				String[] row = rowString.split(",");
 				if (row[0].equals(location) && row[8].length() > 0) {
-					data.add(converter.fromString(row[2]), Integer.parseInt(row[8]));
+					data.add(converter.fromString(row[2]), Float.parseFloat(row[8]));
 				}
 			}
